pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm[toml]>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bci-compression"
dynamic = ["version"]
description = "Neural data compression toolkit for brain-computer interfaces"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Kevin", email = "contact@bci-compression.org" }]
requires-python = ">=3.8"
keywords = ["bci", "compression", "neural-data", "signal-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "torch>=1.13.0",
    "transformers>=4.20.0",
    "scikit-learn>=1.1.0",
    "matplotlib>=3.5.0",
    "h5py>=3.7.0",
    "pywavelets>=1.1.1",
    "pandas>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=2.12.0,<3.0.0",
    "pytest-benchmark>=3.4.0,<4.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "coverage>=6.0.0,<8.0.0",
    "black>=22.0.0,<23.0.0",
    "flake8>=4.0.0,<5.0.0",
    "mypy>=0.991,<1.0.0",
    "pre-commit>=2.17.0,<3.0.0",
    "hypothesis>=6.0.0,<7.0.0",
    "pyyaml>=5.4.0",
    "ruff>=0.0.290",
]
gpu = ["cupy-cuda11x>=12.0.0"]
docs = ["sphinx>=4.5.0,<5.0.0", "sphinx-rtd-theme>=1.0.0,<2.0.0"]
nwb = ["pynwb>=2.5.0", "mne>=1.5.0"]

[project.scripts]
bci-compress = "bci_compression.cli.bci_compress:main"
bci-benchmark = "scripts.benchmark_runner:main"

[project.entry-points."bci_compression.compressors"]
transformer = "bci_compression.algorithms.transformer_compression:TransformerCompressor"
vae = "bci_compression.algorithms.vae_compression:VAECompressor"
spike = "bci_compression.algorithms.spike_codec:SpikeCodec"

[project.urls]
Homepage = "https://github.com/hkevin01/brain-computer-compression"
Documentation = "https://brain-computer-compression.readthedocs.io/"
Repository = "https://github.com/hkevin01/brain-computer-compression"
"Bug Reports" = "https://github.com/hkevin01/brain-computer-compression/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "bci_compression.__version__" }

# Tool configurations
[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU (deselect with '-m \"not gpu\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
